moved BigQueryTargetConfig to pipeline_config
