+++ tests/unit_test/utils/pipeline_config_test.py
@@ -1,9 +1,22 @@
 import pytest
 
 from data_pipeline.utils.pipeline_config import (
+    BigQueryTargetConfig,
     str_to_bool,
     get_environment_variable_value
 )
+
+
+class TestBigQueryTargetConfig:
+    def test_should_read_project_dataset_and_table_name(self):
+        config = BigQueryTargetConfig.from_dict({
+            'projectName': 'project1',
+            'datasetName': 'dataset1',
+            'tableName': 'table1'
+        })
+        assert config.project_name == 'project1'
+        assert config.dataset_name == 'dataset1'
+        assert config.table_name == 'table1'
 
 
 class TestStrToBool:
